% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/manager.R
\name{newConnection}
\alias{newConnection}
\title{Easy connection to a Postgres database using aliases}
\usage{
newConnection(dbAlias, pgPassFile = "~/.pgpass", dbConfFile = "~/db.conf")
}
\arguments{
\item{dbAlias}{alias of the database you want to connect to (must be present in db.conf)}

\item{pgPassFile}{location of ".pgpass" on your filesystem (optional; defaults to "~/.pgpass")}

\item{dbConfFile}{location of "db.conf" on your filesystem (optional; defaults to "~/db.conf")}
}
\value{
PostgresConnectionManager object including a PostgreSQLDriver and a PostgresSQLConnection
         to the database in question
}
\description{
Creates a PostgreSQL driver and a connection to the database \code{dbAlias}
using the .pgpass and db.conf files specified.
}
\details{
The db.conf file is defined one alias per line. Each alias is defined using this format:
\code{#alias:hostname:port:database:username}
\code{database1:my.postgres.db.com:5432:my_database1:user1}
}
\examples{
newConnection("my_db")
newConnection("my_db", dbConfFile = "/path/to/your/db.conf")
}
\seealso{
\code{\link{runQuery}}, \code{\link{destroyConnection}}
}

